requirements.txt
absl-py==2.2.1
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
daphne==4.1.2
debugpy==1.8.13
decorator==5.2.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
executing==2.2.0
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
incremental==24.7.2
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
# protobuf==6.30.2
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
# pywin32==310
PyYAML==6.0.2
pyzmq==26.3.0
redis==5.2.1
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
service-identity==24.2.0
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2025.2.18
# torch==2.6.0
# torchaudio==2.6.0
# torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.0
tzdata==2025.1
ultralytics==8.3.97
ultralytics-thop==2.0.14
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
zope.interface==7.2
absl-py==2.2.1
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3        
certifi==2025.1.31       
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
executing==2.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.5.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
# numpy==2.0.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
protobuf==5.29.4
psutil==7.0.0
psycopg2==2.9.10
pure_eval==0.2.3
py-cpuinfo==9.0.0
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
pywin32==310
PyYAML==6.0.2
pyzmq==26.3.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
rich==14.0.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
# tensorflow_intel==2.18.0
termcolor==3.0.1
threadpoolctl==3.5.0
tifffile==2025.2.18
# torch==2.6.0+cu124
# torchaudio==2.6.0+cu124
# torchvision==0.21.0+cu124
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.0
tzdata==2025.1
ultralytics==8.3.97
ultralytics-thop==2.0.14
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
